DROID-1865 tests
